feat(app): add UserId to post use cases
UserId added to commands and result.
Only owner of the post allowed to update or delete the post.

+++ src/CABasicCRUD.Application/Features/Posts/DeletePost/DeletePostCommandHandler.cs
@@ -1,5 +1,6 @@
 using CABasicCRUD.Application.Common.Interfaces;
 using CABasicCRUD.Application.Common.Interfaces.Messaging;
+using CABasicCRUD.Application.Features.Auth;
 using CABasicCRUD.Domain.Common;
 using CABasicCRUD.Domain.Posts;
 
@@ -21,6 +22,11 @@ internal sealed class DeletePostCommandHandler(
         {
             // throw new KeyNotFoundException($"Post with Id: {request.PostId.Value} not found");
             return Result.Failure(PostErrors.NotFound);
+        }
+
+        if (request.UserId != post.UserId)
+        {
+            return Result.Failure(AuthErrors.Forbidden);
         }
 
         await _postRepository.DeleteAsync(entity: post);
